Norme passed for most files. Add dollar functions to dollar_utils.c
