+++ src/check_dollar.c
