+++ src/check_n.c
