+++ src/envp_cp.c
