--- src/free_data.c
@@ -1,10 +1,0 @@
-#include "../include/minishell.h"
-
-void free_data(t_data *data)
-{
-	data->indexmeta = data->save_indexmeta;
-	free(data->indexmeta);
-	data->line = data->save_line;
-	free(data->line);
-	free_the_pp(data->path_split);
-}
