+++ src/free_memory.c
@@ -19,3 +19,29 @@ void	free_cmd(t_cmd *cmd)
 	}
 	free(cmd);
 }
+
+void	free_the_pp(char **pp)
+{
+	int	i;
+
+	i = 0;
+	if (!pp || !*pp)
+		return ;
+	while (pp[i])
+	{
+		free(pp[i]);
+		pp[i] = NULL;
+		i++;
+	}
+	free(pp);
+	pp = NULL;
+}
+
+void	free_data(t_data *data)
+{
+	data->indexmeta = data->save_indexmeta;
+	free(data->indexmeta);
+	data->line = data->save_line;
+	free(data->line);
+	free_the_pp(data->path_split);
+}
