+++ src/handle_builtin.c
@@ -10,7 +10,6 @@
 
 void	handle_builtin(t_cmd *cmd, t_data *data)
 {
-
 	if (ft_strcmp(cmd->argv[0], "echo") == 0)
 		handle_echo(cmd);
 	else if (ft_strcmp(cmd->argv[0], "cd") == 0)
