+++ src/print_intro.c
@@ -1,35 +1,69 @@
 #include "../include/minishell.h"
 
-// void	print_intro(void)
-// {
-// 	printf("MMMMMMMM               MMMMMMMM    SSSSSSSSSSSSSSS  HHHHHHHHH     HHHHHHHHH\n\
-// M:::::::M             M:::::::M  SS:::::::::::::::S H:::::::H     H:::::::H\n\
-// M::::::::M           M::::::::M S:::::SSSSSS::::::S H:::::::H     H:::::::H\n\
-// M:::::::::M         M:::::::::M S:::::S     SSSSSSS HH::::::H     H::::::HH\n\
-// M::::::::::M       M::::::::::M S:::::S               H:::::H     H:::::H  \n\
-// M:::::::::::M     M:::::::::::M S:::::S               H:::::H     H:::::H  \n\
-// M:::::::M::::M   M::::M:::::::M  S::::SSSS            H::::::HHHHH::::::H  \n\
-// M::::::M M::::M M::::M M::::::M   SS::::::SSSSS       H:::::::::::::::::H  \n\
-// M::::::M  M::::M::::M  M::::::M     SSS::::::::SS     H:::::::::::::::::H  \n\
-// M::::::M   M:::::::M   M::::::M        SSSSSS::::S    H::::::HHHHH::::::H  \n\
-// M::::::M    M:::::M    M::::::M             S:::::S   H:::::H     H:::::H  \n\
-// M::::::M     MMMMM     M::::::M             S:::::S   H:::::H     H:::::H  \n\
-// M::::::M               M::::::M SSSSSSS     S:::::S HH::::::H     H::::::HH\n\
-// M::::::M               M::::::M S::::::SSSSSS:::::S H:::::::H     H:::::::H\n\
-// M::::::M               M::::::M S:::::::::::::::SS  H:::::::H     H:::::::H\n\
-// MMMMMMMM               MMMMMMMM  SSSSSSSSSSSSSSS    HHHHHHHHH     HHHHHHHHH\n");
-// }
-
 void	print_intro(void)
 {
 	printf("\033[0;34m");
-	printf("::::    ::::   ::::::::  :::    :::\n\
-+:+:+: :+:+:+ :+:    :+: :+:    :+:\n\
-+:+ +:+:+ +:+ +:+        +:+    +:+\n\
-+#+  +:+  +#+ +#++:++#++ +#++:++#++\n\
-+#+       +#+        +#+ +#+    +#+\n\
-#+#       #+# #+#    #+# #+#    #+#\n\
-###       ###  ########  ###    ###\n\
-");
-printf("\033[0;35m");
+	printf("MMMMMMMM               MMMMMMMM    SSSSSSSSSSSSSSS  HHHHHHHHH     HHHHHHHHH\n\
+M:::::::M             M:::::::M  SS:::::::::::::::S H:::::::H     H:::::::H\n\
+M::::::::M           M::::::::M S:::::SSSSSS::::::S H:::::::H     H:::::::H\n\
+M:::::::::M         M:::::::::M S:::::S     SSSSSSS HH::::::H     H::::::HH\n\
+M::::::::::M       M::::::::::M S:::::S               H:::::H     H:::::H  \n\
+M:::::::::::M     M:::::::::::M S:::::S               H:::::H     H:::::H  \n\
+M:::::::M::::M   M::::M:::::::M  S::::SSSS            H::::::HHHHH::::::H  \n\
+M::::::M M::::M M::::M M::::::M   SS::::::SSSSS       H:::::::::::::::::H  \n\
+M::::::M  M::::M::::M  M::::::M     SSS::::::::SS     H:::::::::::::::::H  \n\
+M::::::M   M:::::::M   M::::::M        SSSSSS::::S    H::::::HHHHH::::::H  \n\
+M::::::M    M:::::M    M::::::M             S:::::S   H:::::H     H:::::H  \n\
+M::::::M     MMMMM     M::::::M             S:::::S   H:::::H     H:::::H  \n\
+M::::::M               M::::::M SSSSSSS     S:::::S HH::::::H     H::::::HH\n\
+M::::::M               M::::::M S::::::SSSSSS:::::S H:::::::H     H:::::::H\n\
+M::::::M               M::::::M S:::::::::::::::SS  H:::::::H     H:::::::H\n\
+MMMMMMMM               MMMMMMMM  SSSSSSSSSSSSSSS    HHHHHHHHH     HHHHHHHHH\n");
+	printf("\033[0;35m");
 }
+
+void	pepe(void)
+{
+	printf("\033[0;32m");
+	printf("You have been visited by a wild pepe!\n\
+⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⣴⣿⠽⠭⣥⣤⣀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀\n\
+⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⡴⠞⠉⠁⠀⠀⠀⠀⠉⠉⠛⠶⣤⣀⠀⠀⢀⣤⠴⠞⠛⠉⠉⠉⠛⠶⣦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀\n\
+⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⡾⠉⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠉⠳⣏⠉⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⢻⣆⠀⠀⠀⠀⠀⠀⠀⠀\n\
+⠀⠀⠀⠀⠀⠀⠀⠀⠀⣰⠏⠀⠀⠀⠀⠀⠀⢀⣠⠤⠤⠤⠤⢤⣄⡀⠀⠀⠹⡆⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢹⡄⠀⠀⠀⠀⠀⠀⠀\n\
+⠀⠀⠀⠀⠀⠀⠀⢀⡾⠁⠀⠀⠀⠀⠀⠐⠈⠁⠀⠀⠀⠀⠀⠀⠀⠉⠛⠶⢤⣽⡦⠐⠒⠒⠂⠀⠀⠀⠀⠐⠒⠀⢿⣦⣀⡀⠀⠀⠀⠀\n\
+⠀⠀⠀⠀⠀⠀⢀⡞⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⣠⡤⠤⠤⠤⠤⠠⠌⢻⣆⡀⠀⠀⠀⣀⣀⣀⡀⠤⠤⠄⠠⢉⣙⡿⣆⡀⠀\n\
+⠀⠀⠀⠀⣀⣴⡟⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⣤⢶⣛⣩⣶⣶⡾⢯⠿⠷⣖⣦⣤⣍⣿⣴⠖⣋⠭⣷⣶⣶⡶⠒⠒⣶⣒⣠⣀⣙⣿⣆\n\
+⠀⠀⢀⠞⠋⠀⡇⠀⠀⠀⠀⠀⠀⢀⣠⡶⣻⡯⣲⡿⠟⢋⣵⣛⣾⣿⣷⡄⠀⠈⠉⠙⠛⢻⣯⠤⠚⠋⢉⣴⣻⣿⣿⣷⣼⠁⠉⠛⠺⣿\n\
+⠀⣠⠎⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢿⣟⣫⣿⠟⠉⠀⠀⣾⣿⣻⣿⣤⣿⣿⠀⠀⠀⠀⠀⠀⣿⠀⠀⠀⣿⣿⣻⣿⣼⣿⣿⠇⠀⠀⠀⢙\n\
+⢠⠏⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⠛⡶⣄⠀⠀⢻⣿⣿⣿⣿⣿⡏⠀⠀⠀⣀⣤⣾⣁⠀⠀⠀⠸⢿⣿⣿⣿⡿⠋⠀⣀⣠⣶⣿\n\
+⠟⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⠺⢿⣶⣶⣮⣭⣭⣭⣭⡴⢶⣶⣾⠿⠟⠋⠉⠉⠙⠒⠒⠊⠉⠈⠉⠚⠉⠉⢉⣷⡾⠯\n\
+⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠉⠉⠉⠀⠀⠀⢈⣽⠟⠁⠀⠀⠀⠀⣄⡀⠀⠀⠀⠀⠀⠀⢀⣴⡾⠟⠀⠀⠀\n\
+⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⣤⡴⠞⠋⠁⠀⠀⠀⠀⠀⠀⠈⠙⢷⡀⠉⠉⠉⠀⠙⢿⣵⡄⠀⠀⠀\n\
+⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢷⡀⠀⠀\n\
+⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⣀⣀⣀⣀⣀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⣧⠀⠀\n\
+⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⣴⠟⠋⠉⠀⠀⠉⠛⠛⠛⠛⠷⠶⠶⠶⠶⠤⢤⣤⣤⣀⣀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⣠⡤⢿⣆⠀\n\
+⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⡶⠋⠀⠀⠀⠸⠿⠛⠛⠛⠓⠒⠲⠶⢤⣤⣄⣀⠀⠀⠀⠈⠙⠛⠛⠛⠛⠒⠶⠶⠶⣶⠖⠛⠛⠁⢠⣸⡟⠀\n\
+⠀⠀⠀⠀⠀⠀⢰⣆⠀⢸⣧⣤⣤⣤⣤⣤⣤⣤⣤⣤⣀⠀⠀⠀⠀⠀⠉⠉⠛⠛⠓⠒⠲⠦⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣾⠋⠀⠀\n\
+⡀⠀⠀⠀⠀⠀⠀⠙⢷⡄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠉⠉⠉⠛⠲⠶⣶⣤⣄⣀⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⡾⠃⠀⠀⠀\n\
+⣿⣤⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠛⠛⣳⣶⡶⠟⠉⠀⠀⠀⠀⠀\n\
+⠛⢷⣿⣷⠤⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣠⣴⠟⠋⠀⠀⠀⠀⠀⠀⠀⠀⠀\n\
+⠀⠀⠈⠙⠻⢷⣬⣗⣒⣂⡀⠠⠀⠀⣀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣀⣀⣀⣤⡴⠾⠋⠉⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀\n\
+⠀⠀⠀⠀⠀⠀⠀⠉⠉⠛⠛⠿⠶⠶⠶⠶⣤⣤⣭⣭⣍⣉⣉⣀⣀⣀⣀⣼⣯⡽⠷⠿⠛⠙⠿⣦⣀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀\n\
+⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⠉⠉⠉⠉⠉⠉⠉⠉⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢷⡄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀\n\
+⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⡀⠀⠀⠀⠀⠈⠻⣦⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀\n");
+	printf("\033[0;35m");
+}
+
+// void	print_intro(void)
+// {
+// 	printf("\033[0;34m");
+// 	printf("::::    ::::   ::::::::  :::    :::\n\
+// +:+:+: :+:+:+ :+:    :+: :+:    :+:\n\
+// +:+ +:+:+ +:+ +:+        +:+    +:+\n\
+// +#+  +:+  +#+ +#++:++#++ +#++:++#++\n\
+// +#+       +#+        +#+ +#+    +#+\n\
+// #+#       #+# #+#    #+# #+#    #+#\n\
+// ###       ###  ########  ###    ###\n\
+// ");
+// printf("\033[0;35m");
+// }
