+++ src/search_cmds_fd_utils.c
@@ -21,7 +21,6 @@ void	set_fd_out(t_cmd *cmd, int append, t_data *data)
 {
 	if (!cmd || !data->line)
 		return ;
-	//printf("line out: %s\n", data->line);
 	while (*data->line == '\0' && ft_strchr(" >", *data->indexmeta))
 	{
 		data->line++;
@@ -55,7 +54,7 @@ void	get_fd(t_cmd *cmd, t_data *data, char meta)
 
 void	close_fd(t_cmd *cmd)
 {
-	t_cmd *tmp;
+	t_cmd	*tmp;
 
 	tmp = cmd;
 	while (tmp != NULL)
