+++ src/singleton_statuscode.c
@@ -16,13 +16,3 @@ void	set_exit_code(int status_code)
 	code = get_exit_code();
 	*code = status_code;
 }
-//
-// int main()
-// {
-// 	set_exit_code(123);
-// 	int	*exit_code = get_exit_code();
-// 	printf("first: %i", *exit_code);
-// 	set_exit_code(116);
-// 	exit_code = get_exit_code();
-// 	printf("second: %i", *exit_code);
-// }
