+++ src/skip_dollar.c
